Rename FilterCommandParserTest to CourseCommandParserTest
