--- src/test/java/seedu/address/logic/parser/FilterCommandParserTest.java
@@ -1,27 +1,0 @@
-package seedu.address.logic.parser;
-
-import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailure;
-
-import org.junit.jupiter.api.Test;
-
-import seedu.address.logic.commands.CourseCommand;
-
-public class FilterCommandParserTest {
-    private CourseCommandParser parser = new CourseCommandParser();
-
-    @Test
-    public void parse_validArgs_returnsFilterCommand() {
-        // TODO: To be done after FilterCommand and FilterCommandParser is improved
-    }
-
-    @Test
-    public void parse_invalidArgs_throwsParseException() {
-        assertParseFailure(parser, "a",
-                String.format(MESSAGE_INVALID_COMMAND_FORMAT, CourseCommand.MESSAGE_USAGE));
-        assertParseFailure(parser, " ",
-                String.format(MESSAGE_INVALID_COMMAND_FORMAT, CourseCommand.MESSAGE_USAGE));
-        assertParseFailure(parser, "add tg/2",
-                String.format(MESSAGE_INVALID_COMMAND_FORMAT, CourseCommand.MESSAGE_USAGE));
-    }
-}
